Updated function in bls utils
Filled in json_to_df (changed to dict_to_df to accurately represent function). Filled in df_to_csv. Functions may not work for other data - the way this data is formatted is not great for universal calling. Changed list of states to be separate csv file that we can read so it keeps the code cleaner.